backend/model_service/requirements.txt
Flask >= 2.2.2
mysql-connector-python
Pillow
numpy
torch
torchvision 
torchaudio